+++ gd-gltg/CHANGELOG.md
@@ -49,6 +49,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GLGVO-646](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-646)
 - Updated Unmonitored Areas layer (now called Extrapolated Areas)
   [GLGVO-652](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-652)
+- Make rivers show above boundary layers 
+  [GLGVO-658](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-658)
 
 ### Fixed
 - Click the 'x' in the top right corner of the Detail Page, the page returns to the Explore Page